# Add support for multi-argument functions to `function-calc-no-unspaced-operator`