--- a/lib/rules/function-calc-no-unspaced-operator/__tests__/index.mjs
+++ b/lib/rules/function-calc-no-unspaced-operator/__tests__/index.mjs
@@ -195,10 +195,6 @@ testRule({
 			description: 'CRLF newline and tab before operator',
 		},
 		{
-			code: 'a { top: calc(5--6px-7px); }',
-			description: 'custom units',
-		},
-		{
 			code: 'a { padding: 10px calc(calc(1px + 2px)* 3px); }',
 		},
 		{
@@ -249,6 +245,10 @@ testRule({
 		},
 		{
 			code: 'a { padding: 10px calc([10px+5px]); }',
+		},
+		{
+			code: 'a { top: calc(5--6px-7px); }',
+			description: 'custom units',
 		},
 	],