--- a/lib/rules/function-calc-no-unspaced-operator/__tests__/index.mjs
+++ b/lib/rules/function-calc-no-unspaced-operator/__tests__/index.mjs
@@ -225,12 +225,6 @@ testRule({
 			code: 'a { top: calc(10px/var(--foo)); }',
 		},
 		{
-			code: 'a { top: calc(10px*var(--foo, 10px 20px)); }',
-		},
-		{
-			code: 'a { top: calc(10px*var(--foo, 10px+20px)); }',
-		},
-		{
 			code: 'a { padding: calc(); }',
 		},
 		{
@@ -249,6 +243,12 @@ testRule({
 		{
 			code: 'a { top: calc(5--6px-7px); }',
 			description: 'custom units',
+		},
+		{
+			code: 'a { top: calc(10px*var(--foo, 10px 20px)); }',
+		},
+		{
+			code: 'a { top: calc(10px*var(--foo, 10px+20px)); }',
 		},
 	],