--- a/lib/rules/function-calc-no-unspaced-operator/__tests__/index.mjs
+++ b/lib/rules/function-calc-no-unspaced-operator/__tests__/index.mjs
@@ -687,57 +687,57 @@ testRule({
 	],
 });
 
-// testRule({
-// 	ruleName,
-// 	config: [true],
-// 	customSyntax: 'postcss-scss',
-// 	fix: true,
+testRule({
+	ruleName,
+	config: [true],
+	customSyntax: 'postcss-scss',
+	fix: true,
 
-// 	accept: [
-// 		{
-// 			code: 'a { top: calc(100%*#{$foo}); }',
-// 		},
-// 		{
-// 			code: 'a { top: calc(100% *#{$foo}); }',
-// 		},
-// 		{
-// 			code: 'a { top: calc(100%* #{$foo}); }',
-// 		},
-// 		{
-// 			code: 'a { top: calc(100% * #{$foo}); }',
-// 		},
-// 		{
-// 			code: 'a { top: calc(100%/#{$foo}); }',
-// 		},
-// 		{
-// 			code: 'a { top: calc(100% /#{$foo}); }',
-// 		},
-// 		{
-// 			code: 'a { top: calc(100%/ #{$foo}); }',
-// 		},
-// 		{
-// 			code: 'a { top: calc(100% / #{$foo}); }',
-// 		},
-// 	],
+	accept: [
+		{
+			code: 'a { top: calc(100%*#{$foo}); }',
+		},
+		{
+			code: 'a { top: calc(100% *#{$foo}); }',
+		},
+		{
+			code: 'a { top: calc(100%* #{$foo}); }',
+		},
+		{
+			code: 'a { top: calc(100% * #{$foo}); }',
+		},
+		{
+			code: 'a { top: calc(100%/#{$foo}); }',
+		},
+		{
+			code: 'a { top: calc(100% /#{$foo}); }',
+		},
+		{
+			code: 'a { top: calc(100%/ #{$foo}); }',
+		},
+		{
+			code: 'a { top: calc(100% / #{$foo}); }',
+		},
+	],
 
-// 	reject: [
-// 		{
-// 			code: 'a { top: calc(100%- #{$foo}); }',
-// 			fixed: 'a { top: calc(100% - #{$foo}); }',
-// 			message: messages.expectedBefore('-'),
-// 			line: 1,
-// 			column: 19,
-// 			endLine: 1,
-// 			endColumn: 20,
-// 		},
-// 		{
-// 			code: 'a { top: calc(100% -#{$foo}); }',
-// 			fixed: 'a { top: calc(100% - #{$foo}); }',
-// 			message: messages.expectedAfter('-'),
-// 			line: 1,
-// 			column: 20,
-// 			endLine: 1,
-// 			endColumn: 21,
-// 		},
-// 	],
-// });
+	reject: [
+		{
+			code: 'a { top: calc(100%- #{$foo}); }',
+			fixed: 'a { top: calc(100% - #{$foo}); }',
+			message: messages.expectedBefore('-'),
+			line: 1,
+			column: 19,
+			endLine: 1,
+			endColumn: 20,
+		},
+		{
+			code: 'a { top: calc(100% -#{$foo}); }',
+			fixed: 'a { top: calc(100% - #{$foo}); }',
+			message: messages.expectedAfter('-'),
+			line: 1,
+			column: 20,
+			endLine: 1,
+			endColumn: 21,
+		},
+	],
+});