--- a/lib/rules/function-calc-no-unspaced-operator/__tests__/index.mjs
+++ b/lib/rules/function-calc-no-unspaced-operator/__tests__/index.mjs
@@ -239,6 +239,11 @@ testRule({
 		},
 		{
 			code: 'a { padding: 10px calc([10px+5px]); }',
+			description: 'simple blocks with square brackets are ignored',
+		},
+		{
+			code: 'a { padding: 10px calc({10px+5px}); }',
+			description: 'simple blocks with curly braces are ignored',
 		},
 		{
 			code: 'a { top: calc(5--6px-7px); }',
@@ -620,6 +625,16 @@ testRule({
 			column: 19,
 			endLine: 1,
 			endColumn: 20,
+		},
+		{
+			code: 'a { top: calc(calc(1px+ 1px)); }',
+			fixed: 'a { top: calc(calc(1px + 1px)); }',
+			description: 'nested math expression',
+			message: messages.expectedBefore('+'),
+			line: 1,
+			column: 23,
+			endLine: 1,
+			endColumn: 24,
 		},
 		{
 			code: 'a { top: calc(((1px+ 1px))); }',