--- a/lib/rules/function-calc-no-unspaced-operator/index.cjs
+++ b/lib/rules/function-calc-no-unspaced-operator/index.cjs
@@ -179,7 +179,7 @@ const rule = (primary, _secondaryOptions, context) => {
 
 			/**
 			 * @param {import('@csstools/css-parser-algorithms').ContainerNode} node
-			 * @param {OperationWithoutOperator} operation
+			 * @param {Operation} operation
 			 */
 			function checkOperationWithoutOperator(node, operation) {
 				if (
@@ -350,11 +350,7 @@ const rule = (primary, _secondaryOptions, context) => {
 							checkOperandWhitespace(operation, operation.after, 'after');
 						}
 
-						if (operation.secondOperand) {
-							cursor = node.value.indexOf(operation.secondOperand);
-						} else {
-							cursor++;
-						}
+						cursor = node.value.indexOf(operation.secondOperand);
 					}
 				},
 				{
@@ -401,9 +397,9 @@ function isOperandNode(node) {
  * NOTE: this is messy.
  *
  * @typedef {{
- *   firstOperand: import ('@csstools/css-parser-algorithms').ComponentValue | undefined,
+ *   firstOperand: import ('@csstools/css-parser-algorithms').ComponentValue,
  *   before: Array<import ('@csstools/css-parser-algorithms').ComponentValue>,
- *   secondOperand: import ('@csstools/css-parser-algorithms').ComponentValue | undefined,
+ *   secondOperand: import ('@csstools/css-parser-algorithms').ComponentValue,
  *   after: Array<import ('@csstools/css-parser-algorithms').ComponentValue>,
  *   operator: import ('@csstools/css-parser-algorithms').ComponentValue | undefined,
  *   operatorToken: import ('@csstools/css-tokenizer').TokenDelim | undefined,
@@ -417,15 +413,6 @@ function isOperandNode(node) {
  *   operator: import ('@csstools/css-parser-algorithms').ComponentValue,
  *   operatorToken: import ('@csstools/css-tokenizer').TokenDelim,
  * }} CompleteOperation
- *
- * @typedef {{
- *   firstOperand: import ('@csstools/css-parser-algorithms').ComponentValue,
- *   before: Array<import ('@csstools/css-parser-algorithms').ComponentValue>,
- *   secondOperand: import ('@csstools/css-parser-algorithms').ComponentValue,
- *   after: Array<import ('@csstools/css-parser-algorithms').ComponentValue>,
- *   operator: import ('@csstools/css-parser-algorithms').ComponentValue | undefined,
- *   operatorToken: import ('@csstools/css-tokenizer').TokenDelim | undefined,
- * }} OperationWithoutOperator
  */
 
 /**
@@ -443,7 +430,7 @@ function isCompleteOperation(operation) {
 
 /**
  * @param {Operation} operation
- * @returns {operation is OperationWithoutOperator}
+ * @returns {operation is Operation}
  */
 function isOperationWithoutOperator(operation) {
 	return (
@@ -460,62 +447,77 @@ function isOperationWithoutOperator(operation) {
  * @returns {[Operation | undefined, number]}
  */
 function parseOperation(container, cursor) {
-	/** @type {Operation} */
-	const operation = {
-		firstOperand: undefined,
-		before: [],
-		secondOperand: undefined,
-		after: [],
-		operator: undefined,
-		operatorToken: undefined,
-	};
+	/** @type {import ('@csstools/css-parser-algorithms').ComponentValue | undefined} */
+	let firstOperand = undefined;
+	/** @type {import ('@csstools/css-parser-algorithms').ComponentValue | undefined} */
+	let secondOperand = undefined;
+	/** @type {Array<import ('@csstools/css-parser-algorithms').ComponentValue>} */
+	const before = [];
+	/** @type {Array<import ('@csstools/css-parser-algorithms').ComponentValue>} */
+	const after = [];
+	/** @type {import ('@csstools/css-parser-algorithms').ComponentValue | undefined} */
+	let operator = undefined;
+	/** @type {import ('@csstools/css-tokenizer').TokenDelim | undefined} */
+	let operatorToken = undefined;
 
 	let currentNode = container.value[cursor];
 
+	// Consume as much whitespace and comments as possible
 	while (cssParserAlgorithms.isWhitespaceNode(currentNode) || cssParserAlgorithms.isCommentNode(currentNode)) {
 		currentNode = container.value[++cursor];
 	}
 
+	// If the current node is an operand, consume it
 	if (isOperandNode(currentNode)) {
-		operation.firstOperand = currentNode;
+		firstOperand = currentNode;
 
 		currentNode = container.value[++cursor];
 	}
 
+	// Consume as much whitespace and comments as possible
+	// Assign to `before`
 	while (cssParserAlgorithms.isWhitespaceNode(currentNode) || cssParserAlgorithms.isCommentNode(currentNode)) {
-		operation.before.push(currentNode);
+		before.push(currentNode);
 
 		currentNode = container.value[++cursor];
 	}
 
+	// If the current node is an operator, consume it
 	if (
 		cssParserAlgorithms.isTokenNode(currentNode) &&
 		currentNode.value[0] === cssTokenizer.TokenType.Delim &&
 		OPERATORS.has(currentNode.value[4].value)
 	) {
-		operation.operator = currentNode;
-		operation.operatorToken = currentNode.value;
+		operator = currentNode;
+		operatorToken = currentNode.value;
 
 		currentNode = container.value[++cursor];
 	}
 
+	// Consume as much whitespace and comments as possible
+	// Assign to `after`
 	while (cssParserAlgorithms.isWhitespaceNode(currentNode) || cssParserAlgorithms.isCommentNode(currentNode)) {
-		operation.after.push(currentNode);
+		after.push(currentNode);
 
 		currentNode = container.value[++cursor];
 	}
 
+	// If the current node is an operand, consume it
 	if (isOperandNode(currentNode)) {
-		operation.secondOperand = currentNode;
+		secondOperand = currentNode;
 
 		currentNode = container.value[++cursor];
 	}
 
+	// Consume as much whitespace and comments as possible
 	while (cssParserAlgorithms.isWhitespaceNode(currentNode) || cssParserAlgorithms.isCommentNode(currentNode)) {
 		currentNode = container.value[++cursor];
 	}
 
-	if (!operation.firstOperand || !operation.secondOperand) {
+	// If we have not consumed any operands, we are not in an operation
+	// Do error recovery by consuming until the next comma or semicolon
+	// If no comma or semicolon is found, consume until the end of the container
+	if (!firstOperand || !secondOperand) {
 		while (currentNode) {
 			if (
 				cssParserAlgorithms.isTokenNode(currentNode) &&
@@ -530,7 +532,17 @@ function parseOperation(container, cursor) {
 		return [undefined, container.value.length];
 	}
 
-	return [operation, cursor];
+	return [
+		{
+			firstOperand,
+			before,
+			secondOperand,
+			after,
+			operator,
+			operatorToken,
+		},
+		cursor,
+	];
 }
 
 rule.ruleName = ruleName;