--- a/lib/rules/function-calc-no-unspaced-operator/index.cjs
+++ b/lib/rules/function-calc-no-unspaced-operator/index.cjs
@@ -140,7 +140,7 @@ const rule = (primary, _secondaryOptions, context) => {
 			function checkCompleteOperation(node, operation) {
 				if (operation.before.length && operation.after.length) return;
 
-				if (!operation.before.length) {
+				if (!operation.before.find(cssParserAlgorithms.isWhitespaceNode)) {
 					if (context.fix) {
 						needsFix = true;
 						node.value.splice(
@@ -158,7 +158,7 @@ const rule = (primary, _secondaryOptions, context) => {
 					}
 				}
 
-				if (!operation.after.length) {
+				if (!operation.after.find(cssParserAlgorithms.isWhitespaceNode)) {
 					if (context.fix) {
 						needsFix = true;
 						node.value.splice(
@@ -255,10 +255,10 @@ const rule = (primary, _secondaryOptions, context) => {
 			 * @param {'before' | 'after'} position
 			 */
 			function checkOperandWhitespace(operation, whitespaceNodes, position) {
-				const firstWhitespaceNode = whitespaceNodes[0];
+				whitespaceNodes.forEach((whitespaceNode) => {
+					if (!cssParserAlgorithms.isWhitespaceNode(whitespaceNode)) return;
 
-				if (whitespaceNodes.length === 1 && cssParserAlgorithms.isWhitespaceNode(firstWhitespaceNode)) {
-					const whitespace = firstWhitespaceNode.toString();
+					const whitespace = whitespaceNode.toString();
 
 					if (whitespace === ' ') return;
 
@@ -269,7 +269,7 @@ const rule = (primary, _secondaryOptions, context) => {
 					if (context.fix) {
 						needsFix = true;
 
-						firstWhitespaceNode.value = [
+						whitespaceNode.value = [
 							[
 								cssTokenizer.TokenType.Whitespace,
 								indexOfFirstNewLine === -1 ? ' ' : whitespace.slice(indexOfFirstNewLine),
@@ -289,7 +289,7 @@ const rule = (primary, _secondaryOptions, context) => {
 							operation.operatorToken[4].value,
 						);
 					}
-				}
+				});
 			}
 
 			cssParserAlgorithms.walk(