--- a/lib/rules/function-calc-no-unspaced-operator/index.cjs
+++ b/lib/rules/function-calc-no-unspaced-operator/index.cjs
@@ -95,7 +95,7 @@ const rule = (primary, _secondaryOptions, context) => {
 					}`;
 					token[3] = token[2] + token[1].length;
 
-					const remainderTokens = cssTokenizer.tokenize({ css: remainder }).slice(0, -1);
+					const remainderTokens = cssTokenizer.tokenize({ css: remainder }).slice(0, -1); // Trim EOF token
 
 					remainderTokens.forEach((remainderToken) => {
 						remainderToken[2] += token[3];