--- a/lib/rules/function-calc-no-unspaced-operator/index.mjs
+++ b/lib/rules/function-calc-no-unspaced-operator/index.mjs
@@ -148,7 +148,7 @@ const rule = (primary, _secondaryOptions, context) => {
 			function checkCompleteOperation(node, operation) {
 				if (operation.before.length && operation.after.length) return;
 
-				if (!operation.before.length) {
+				if (!operation.before.find(isWhitespaceNode)) {
 					if (context.fix) {
 						needsFix = true;
 						node.value.splice(
@@ -166,7 +166,7 @@ const rule = (primary, _secondaryOptions, context) => {
 					}
 				}
 
-				if (!operation.after.length) {
+				if (!operation.after.find(isWhitespaceNode)) {
 					if (context.fix) {
 						needsFix = true;
 						node.value.splice(
@@ -263,10 +263,10 @@ const rule = (primary, _secondaryOptions, context) => {
 			 * @param {'before' | 'after'} position
 			 */
 			function checkOperandWhitespace(operation, whitespaceNodes, position) {
-				const firstWhitespaceNode = whitespaceNodes[0];
+				whitespaceNodes.forEach((whitespaceNode) => {
+					if (!isWhitespaceNode(whitespaceNode)) return;
 
-				if (whitespaceNodes.length === 1 && isWhitespaceNode(firstWhitespaceNode)) {
-					const whitespace = firstWhitespaceNode.toString();
+					const whitespace = whitespaceNode.toString();
 
 					if (whitespace === ' ') return;
 
@@ -277,7 +277,7 @@ const rule = (primary, _secondaryOptions, context) => {
 					if (context.fix) {
 						needsFix = true;
 
-						firstWhitespaceNode.value = [
+						whitespaceNode.value = [
 							[
 								TokenType.Whitespace,
 								indexOfFirstNewLine === -1 ? ' ' : whitespace.slice(indexOfFirstNewLine),
@@ -297,7 +297,7 @@ const rule = (primary, _secondaryOptions, context) => {
 							operation.operatorToken[4].value,
 						);
 					}
-				}
+				});
 			}
 
 			walk(