--- a/lib/rules/function-calc-no-unspaced-operator/index.mjs
+++ b/lib/rules/function-calc-no-unspaced-operator/index.mjs
@@ -103,7 +103,7 @@ const rule = (primary, _secondaryOptions, context) => {
 					}`;
 					token[3] = token[2] + token[1].length;
 
-					const remainderTokens = tokenize({ css: remainder }).slice(0, -1);
+					const remainderTokens = tokenize({ css: remainder }).slice(0, -1); // Trim EOF token
 
 					remainderTokens.forEach((remainderToken) => {
 						remainderToken[2] += token[3];