> Which issue, if any, is this issue related to?

Closes #7618

<hr />

@ybiquitous This rule and the implementation was a lot more complex that I originally thought it would be. Doing pure code review was tricky for me and I needed to take a deep dive.

_The code in this PR is a mess, but it also isn't intended to be used as is.
It can be used, we can continue work from this. But I haven't spend much time on self-review to get it into a really good state._

As I understand it know there are multiple issues and layers.

Steps:
1. Fix units : `10px-20px` -> `10px -20px` (`px-20px` is the entire unit here)
2. Fix missing operators : `10px -20px` -> `10px -|20px` (`|` to indicate that these are separate)
3. Insert whitespace : `10px -|20px` -> `10px - 20px`
4. Normalize whitespace : `10px -\t20px` -> `10px - 20px`


The csstools tokenizer and parsing packages have several aspects that help with some of these steps.

The tokenizer is just arrays, strings and numbers. So Step 1 can be handled purely on the token stream. It is a messy function, but it gives us more control and nice separation of concerns.

The parsing algorithm walker can be stateful. This is useful to keep track of being in a math context or not e.g. `calc((((10px+2px))))`

In general I try to do as much work and analysis on the parsed values (e.g. `token[4].value`, `token[4].unit`, ...) and as little as possible on the string representation (e.g. `token[1]`, `node.toString()`)

My hope is that doing the same work from different angles is a good way to transfer some knowledge :)

Ref #7655