@Mouvedia 2024-04-29
If you can't make a benchmark, do you feel confident that your new approach is faster or at least as fast as the current one?