@romainmenke 2024-04-29
> If you can't make a benchmark, do you feel confident that your new approach is faster or at least as fast as the current one?

This change:

```
;node benchmark-rule.mjs function-calc-no-unspaced-operator 'true'
Warnings: 0
Mean: 92.64952612499998 ms
Deviation: 11.291622367245036 ms
```

Main :

```
;node benchmark-rule.mjs function-calc-no-unspaced-operator 'true'
Warnings: 0
Mean: 78.32598956249998 ms
Deviation: 8.242243635683067 ms
```

`function-calc-no-unspaced-operator-bugfix` :

```
;node benchmark-rule.mjs function-calc-no-unspaced-operator 'true'
Warnings: 0
Mean: 96.32743500000002 ms
Deviation: 9.753154793016495 ms
```

These results match my expectations.
This rule is doing more work than before, so it will be slightly slower.