@Mouvedia 2024-04-29
> so it will be _slightly_ slower

Great.  
I am sure we will find some ways to gain some ms during the review.

> My hope is that doing the same work from different angles is a good way to transfer some knowledge :)

I think that once the other PR is merged, you should fix the conflicts and continue your effort.