@Mouvedia 2024-04-29
I didn't check the code yet :)
Personally I don't care about the approach as long as it passes the current tests and if we are gaining something while we are losing on the perf side. The tests are looking good.
The only thing I would like is for both PRs to be part of the same version.

Feel free to switch from draft to ready if you want a proper review.