@romainmenke 2024-05-01
@ybiquitous I've made a few more edits:
- bring back support for `scss` interpolation
- improve coverage
- improve support for comments (e.g. `calc(1px/* a comment */+ 2px)`)

Feel free to take this!