@Mouvedia 2024-05-01
Can you remove the `export` from https://github.com/stylelint/stylelint/blob/f87f784a429f626e67710ab1a8026c0725414a3b/lib/reference/functions.mjs#L29?