@romainmenke 2024-05-01
I've made a few edits, but will defer to @ybiquitous now for the remainder of the work on this rule :)